requirements.txt
Flask
pandas
requests
python-dotenv
pytz
google-generativeai
